+++ R/shp_to_inp.R
@@ -212,18 +212,18 @@ shp_to_inp <- function(path_options = NULL,
     warning("Define path to point file including filename and ending otherwise sections junctions and coordinates are missing.")
 
     # specify object junction which is called when testing column names for warn message in junction_parameters
-    junction <- NULL
+    junctions <- NULL
   } else {
     # read junction point file
-    junction <- sf::st_read(path_point, stringsAsFactors = F, quiet = TRUE) %>%
+    junctions <- sf::st_read(path_point, stringsAsFactors = F, quiet = TRUE) %>%
       tibble::as_tibble() %>%
       compare_to_dictionary()
 
     # check column names:
-    if (all(c("Name", "Bottom") %in% colnames(junction))) {
-      if ("Top" %in% colnames(junction) | "Ymax" %in% colnames(junction)) {
-        list_of_sections[["junction"]] <- junction
-        list_of_sections[["coordinates"]] <- junction
+    if (all(c("Name", "Bottom") %in% colnames(junctions))) {
+      if ("Top" %in% colnames(junctions) | "Ymax" %in% colnames(junctions)) {
+        list_of_sections[["junctions"]] <- junctions
+        list_of_sections[["coordinates"]] <- junctions
       }
     } else {
       stop("The point file has to include at least the columns named: Name, Bottom and Top or Ymax.")
@@ -232,7 +232,7 @@ shp_to_inp <- function(path_options = NULL,
 
 
   if (is.null(junction_parameters)) {
-    if (!("Y" %in% colnames(junction)) | !("Ysur" %in% colnames(junction)) | !("Apond" %in% colnames(junction))) {
+    if (!("Y" %in% colnames(junctions)) | !("Ysur" %in% colnames(junctions)) | !("Apond" %in% colnames(junctions))) {
       warning(" Y, Ysur or Apond are not defined in point.shp or junction_parameters. Check point.shp for completeness otherwise missing parameters in the section junctions will be filled with default values.")
     }
   }
@@ -374,7 +374,7 @@ shp_to_inp <- function(path_options = NULL,
   # adjust order of sections
   section_order <-  c("title", "options", "evaporation", "raingages", "subcatchments", 
                       "subareas", "infiltration", "aquifers", "groundwater", 
-                      "LID_controls", "LID_usage", "junction", "outfalls", "storage", 
+                      "LID_controls", "LID_usage", "junctions", "outfalls", "storage", 
                       "conduits", "pumps", "weirs", "xsections", "controls", "DWF", 
                       "pollutants", "landuses", "coverages", "loadings", "buildup", 
                       "washoff", "inflows", "timeseries", "curves", "patterns",